breaking sigmap functions into separate library for better use across tools
