+++ include/CoreGen/CoreGenSigMap/CMakeLists.txt
@@ -1,0 +1,13 @@
+# include/CoreGen/CoreGenSigMap CMakeLists.txt
+#
+# Copyright (C) 2017-2019 Tactical Computing Laboratories, LLC
+# All Rights Reserved
+# contact@tactcomplabs.com
+#
+# See LICENSE in the top level directory for licensing details
+#
+
+install(DIRECTORY "."
+        DESTINATION "include/CoreGenSigMap"
+        FILES_MATCHING
+        PATTERN "*.h")
