+++ src/CMakeLists.txt
@@ -10,6 +10,7 @@
 add_subdirectory(CoreGenBackend)
 add_subdirectory(CoreGenPlugin)
 add_subdirectory(CoreGenCodegen)
+add_subdirectory(CoreGenSigMap)
 add_subdirectory(StoneCutter)
 add_subdirectory(SCComp)
 add_subdirectory(CGCLI)
